feat: 사용자가 입력한 단어를 서버로 보내는 word route 추가
